S01-T07

See merge request Jonny__/jborn_web_10_dzhenenko!15

+++ src/main/java/ru/dzhenenko/dao/AccountDao.java
@@ -41,7 +41,7 @@ public class AccountDao {
         }
     }
 
-    public AccountModel deleteAccount(int id) {
+    public AccountModel deleteAccount(long id) {
         AccountModel accountModel = new AccountModel();
         try (Connection conn = dataSource.getConnection()) {
             PreparedStatement st = conn.prepareStatement("delete from account where id = ?", Statement.RETURN_GENERATED_KEYS);
